All sections completed
Completed all parts, appears to be functioning 100%, should test again
though.

Fixed a typo in Pin.java that caused an exception when adding to a
design window.

Removed test Main from Switch.java as indicated in comment. Also added a
getter so that other components may access the fields in the
visual-circuit sections.

There is an existing issue where lines do not display between certain
components on the circuit board. This seems to only happen when joining
an LED or Switch; gates, bitadders, etc., all see the lines rendered
fine.